Further minor edits for warnings. There is still one DeprecationWarning that seems to arise based on the interaction between the ensemble calibration and the spot database tests.

+++ lib/improver/tests/ensemble_copula_coupling/ensemble_copula_coupling/test_ResamplePercentiles.py
@@ -105,9 +105,11 @@ class Test__add_bounds_to_percentiles_and_forecast_values(IrisTest):
         forecast_at_percentiles = cube.data.reshape(3, 9)
         bounds_pairing = (-40, 50)
         lower_array = np.full(
-            forecast_at_percentiles[:, 0].shape, bounds_pairing[0])
+            forecast_at_percentiles[:, 0].shape, bounds_pairing[0],
+            dtype=np.int32)
         upper_array = np.full(
-            forecast_at_percentiles[:, 0].shape, bounds_pairing[1])
+            forecast_at_percentiles[:, 0].shape, bounds_pairing[1],
+            dtype=np.int32)
         plugin = Plugin()
         result = plugin._add_bounds_to_percentiles_and_forecast_at_percentiles(
             percentiles, forecast_at_percentiles, bounds_pairing)
